refactor(e18e): split scanner.rs into statements and expressions modules
